Refactor Wordle Form Input
