+++ apps/project-wordle/src/components/FormInput/index.js
@@ -1,0 +1,2 @@
+export * from './FormInput';
+export { default } from './FormInput';
