+++ apps/project-wordle/src/components/Keyboard/Keyboard.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function Keyboard() {
+  return <div></div>;
+}
+
+export default Keyboard;
