+++ apps/project-wordle/src/components/Keyboard/index.js
@@ -1,0 +1,2 @@
+export * from './Keyboard';
+export { default } from './Keyboard';
